PlayerFlight의 방향키 변경 로직 추가

+++ ShootingGame/Point.h
@@ -7,8 +7,8 @@ public:
 
 	double GetX() { return x; }
 	double GetY() { return y; }
-	void SetX(double putx) { x = putx; }
-	void SetY(double puty) { y = puty; }
+	void SetX(float putx) { x = putx; }
+	void SetY(float puty) { y = puty; }
 
 private:
 	float x, y;
